requirements.txt
flask
faiss-cpu
sentence-transformers
openai
python-dotenv
